migrate test_serverdensity_device to pytest
